Disabling the cache at config level for all tests apart from a cache-specific test. Corrections based on code review
