+++ .circleci/config.yml
@@ -50,8 +50,7 @@ executors:
     docker:
       - image: cimg/openjdk:21.0
         environment:
-          _JAVA_OPTIONS: -Xmx1024m -XX:ParallelGCThreads=2 -XX:ConcGCThreads=2
-            -XX:ParallelGCThreads=2
+          _JAVA_OPTIONS: -Xmx1024m -XX:ParallelGCThreads=2 -XX:ConcGCThreads=2            
             -Djava.util.concurrent.ForkJoinPool.common.parallelism=2
             -Dorg.gradle.daemon=false
       - image: postgres:14
