+++ src/test/kotlin/uk/gov/justice/digital/hmpps/makerecalldecisionapi/integration/controller/RecommendationControllerTest.kt
@@ -11,16 +11,12 @@ import org.joda.time.LocalDateTime
 import org.json.JSONArray
 import org.json.JSONObject
 import org.junit.jupiter.api.Assertions.assertNull
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.cache.CacheManager
 import org.springframework.cache.get
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.web.reactive.function.BodyInserters
-import uk.gov.justice.digital.hmpps.makerecalldecisionapi.config.CacheConstants.USER_ACCESS_CACHE_KEY
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.domain.makerecalldecisions.recommendation.bookRecallToPpud
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.domain.makerecalldecisions.recommendation.toJsonString
 import uk.gov.justice.digital.hmpps.makerecalldecisionapi.integration.IntegrationTestBase
@@ -46,14 +42,6 @@ import java.util.UUID
 @ActiveProfiles("test")
 @ExperimentalCoroutinesApi
 class RecommendationControllerTest() : IntegrationTestBase() {
-
-  @Autowired
-  lateinit var cacheManager: CacheManager
-
-  @BeforeEach
-  fun setup() {
-    cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
-  }
 
   @Test
   fun `get latest complete recommendation overview`() {
@@ -233,8 +221,6 @@ class RecommendationControllerTest() : IntegrationTestBase() {
   fun `create recommendation with recallConsideredList feature flag active`() {
     userAccessAllowed(crn)
     personalDetailsResponseOneTimeOnly(crn)
-    // Clear previous cache write
-    cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
     deleteAndCreateRecommendation("{\"flagConsiderRecall\": true, \"unknownFeatureFlag\": true }")
 
     val response = convertResponseToJSONObject(
@@ -346,8 +332,6 @@ class RecommendationControllerTest() : IntegrationTestBase() {
     personalDetailsResponseOneTimeOnly(crn)
     licenceConditionsResponse(crn, 2500614567)
     oasysAssessmentsResponse(crn)
-    // Clear previous cache write
-    cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
     deleteAndCreateRecommendation()
 
     webTestClient.patch()
@@ -1074,11 +1058,7 @@ class RecommendationControllerTest() : IntegrationTestBase() {
     runTest {
       userAccessAllowedOnce(crn)
       personalDetailsResponse(crn)
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       deleteAndCreateRecommendation()
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       userAccessExcluded(crn)
       webTestClient.get()
         .uri("/recommendations/$createdRecommendationId")
@@ -1098,8 +1078,6 @@ class RecommendationControllerTest() : IntegrationTestBase() {
   fun `given case is excluded when creating a recommendation then only return user access details`() {
     runTest {
       userAccessExcluded(crn)
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       webTestClient.post()
         .uri("/recommendations")
         .contentType(MediaType.APPLICATION_JSON)
@@ -1144,8 +1122,6 @@ class RecommendationControllerTest() : IntegrationTestBase() {
     runTest {
       userAccessAllowed(crn)
       personalDetailsResponse(crn)
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       deleteAndCreateRecommendation()
 
       webTestClient.patch()
@@ -1166,11 +1142,7 @@ class RecommendationControllerTest() : IntegrationTestBase() {
       userResponse("some_user", "test@digital.justice.gov.uk")
       userAccessAllowedOnce(crn)
       personalDetailsResponse(crn)
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       deleteAndCreateRecommendation()
-      // Clear previous cache write
-      cacheManager[USER_ACCESS_CACHE_KEY]?.invalidate()
       userAccessExcluded(crn)
       webTestClient.post()
         .uri("/recommendations/$createdRecommendationId/part-a")
