+++ src/test/resources/application-test.yml
@@ -7,6 +7,9 @@ spring:
   flyway:
     enabled: false
 
+  cache:
+    type: none
+
 server:
   shutdown: immediate
